+++ backend/problems/src/ma2.rs
@@ -1,1 +1,2 @@
 mod quadratic_equations;
+mod rules_of_squares;